requirements.txt
apache-beam[gcp]>=2.60.0
google-cloud-monitoring>=2.22.2
protobuf>=4.21.6
